+++ src/Sulu/Component/Content/Types/TextEditor.php
@@ -11,6 +11,7 @@
 
 namespace Sulu\Component\Content\Types;
 
+use Sulu\Component\Content\Compat\PropertyInterface;
 use Sulu\Component\Content\Compat\PropertyParameter;
 use Sulu\Component\Content\SimpleContentType;
 
@@ -19,6 +20,9 @@ use Sulu\Component\Content\SimpleContentType;
  */
 class TextEditor extends SimpleContentType
 {
+    /**
+     * @var string
+     */
     private $template;
 
     public function __construct($template)
@@ -29,9 +33,7 @@ class TextEditor extends SimpleContentType
     }
 
     /**
-     * returns a template to render a form.
-     *
-     * @return string
+     * {@inheritdoc}
      */
     public function getTemplate()
     {
@@ -39,11 +41,9 @@ class TextEditor extends SimpleContentType
     }
 
     /**
-     * returns default parameters.
-     *
-     * @return array
+     * {@inheritdoc}
      */
-    public function getDefaultParams()
+    public function getDefaultParams(PropertyInterface $property = null)
     {
         return [
             'table' => new PropertyParameter('table', true),